Support private repos / organisation membership. Closes #4
The problem was that when a user was a member of an organisation with only private
repositories the script did not have enough privileges to see the organisation membership
or the private repository access.

In order to cover both problems the initializer now generates a token with the additional
access scope "repo", which gives the script access to private repositories in order to
manage their notification subscriptions.
This additional scope implicitly also gives the script read access to the
organisation memberships, allowing it to manage all repository subscriptions properly.